requirements.txt
google-cloud-aiplatform
flask=3.0.3
python-dotenv=0.21.0